==> Dollar General/3. Drop duplicates.ipynb <==
import pandas as pd

!ls

df = pd.read_csv("2_final_df.csv")
df

Current_Penalty_df = df[df['Violations/Penalties'] == 'Current Penalty']
Current_Penalty_df = Current_Penalty_df.drop_duplicates(['Activity'])
Current_Penalty_df

Initial_Penalty_df = df[df['Violations/Penalties'] == 'Initial Penalty']
Initial_Penalty_df = Initial_Penalty_df.drop_duplicates(['Activity'])
Initial_Penalty_df

Current_Violations_df = df[df['Violations/Penalties'] == 'Current Violations']
Current_Violations_df = Current_Violations_df.drop_duplicates(['Activity'])
Current_Violations_df

Initial_Violations_df = df[df['Violations/Penalties'] == 'Initial Violations']
Initial_Violations_df = Initial_Violations_df.drop_duplicates(['Activity'])
Initial_Violations_df

Current_Penalty_df.to_csv('3_currentpenaltydrop.csv', index=False)

Initial_Penalty_df.to_csv('3_initialpenaltydrop.csv', index=False)

Current_Violations_df.to_csv('3_currentviolationsdrop.csv', index=False)

Initial_Violations_df.to_csv('3_initialviolationsdrop.csv', index=False)
